Move whitelist tests into spi
