+++ modules/lang-painless/spi/src/test/resources/org/elasticsearch/painless/spi/org.elasticsearch.painless.annotation.unknown_with_options
@@ -1,5 +1,5 @@
 # whitelist for annotation tests with unknown annotation containing options
 
-class org.elasticsearch.painless.AnnotationTestObject @no_import {
+class org.elasticsearch.painless.spi.AnnotationTestObject @no_import {
   void unknownAnnotationMethod() @unknownAnootationWithMessage[arg="arg value"]
 }
