+++ modules/lang-painless/src/test/java/org/elasticsearch/painless/ThisTests.java
@@ -78,7 +78,7 @@ public class ThisTests extends ScriptTestCase {
     protected Map<ScriptContext<?>, List<Whitelist>> scriptContexts() {
         Map<ScriptContext<?>, List<Whitelist>> contexts = new HashMap<>();
         List<Whitelist> whitelists = new ArrayList<>(PainlessPlugin.BASE_WHITELISTS);
-        whitelists.add(WhitelistLoader.loadFromResourceFiles(Whitelist.class, "org.elasticsearch.painless.this"));
+        whitelists.add(WhitelistLoader.loadFromResourceFiles(PainlessPlugin.class, "org.elasticsearch.painless.this"));
         contexts.put(ThisScript.CONTEXT, whitelists);
         return contexts;
     }
